+++ packages/@expo/metro-runtime/src/promiseRejectionTracking.ts
@@ -1,0 +1,2 @@
+// NOTE(@krystofwoldrich): This is only needed for native runtimes (hermes)
+export function enablePromiseRejectionTracking() {}