+++ packages/expo-app-integrity/android/src/main/java/expo/modules/integrity/IntegrityExceptions.kt
@@ -2,8 +2,9 @@ package expo.modules.integrity
 
 import expo.modules.kotlin.exception.CodedException
 
-internal class IntegrityException(reason: String, cause: Throwable? = null) :
+internal class IntegrityException(code: String, reason: String, cause: Throwable? = null) :
   CodedException(
+    code,
     "[ExpoAppIntegrity]: $reason",
     cause
   )