+++ packages/expo-build-properties/src/pluginConfig.ts
@@ -390,11 +390,12 @@ export interface PluginConfigTypeIos {
   buildReactNativeFromSource?: boolean;
 
   /**
-   * Enables support for prebuilt React Native iOS dependencies (`ReactNativeDependencies.xcframework`).
-   * This feature is available from React Native 0.80 and later.
-   * @deprecated Use `buildReactNativeFromSource` instead.
+   * The React Native release level to use for the project.
+   * This can be used to enable different sets of internal React Native feature flags.
+   *
+   * @default 'stable'
    */
-  buildFromSource?: boolean;
+  reactNativeReleaseLevel?: 'stable' | 'canary' | 'experimental';
 }
 
 /**
@@ -749,7 +750,11 @@ const schema: JSONSchemaType<PluginConfigType> = {
           nullable: true,
         },
         buildReactNativeFromSource: { type: 'boolean', nullable: true },
-        buildFromSource: { type: 'boolean', nullable: true },
+        reactNativeReleaseLevel: {
+          type: 'string',
+          enum: ['stable', 'canary', 'experimental'],
+          nullable: true,
+        },
       },
       nullable: true,
     },