+++ packages/expo-dev-launcher/ios/SwiftUI/SettingsTabView.swift
@@ -68,9 +68,7 @@ struct SettingsTabView: View {
       Toggle("Show menu at launch", isOn: $viewModel.showOnLaunch)
     }
     .padding()
-#if !os(tvOS)
-    .background(Color(.systemGroupedBackground))
-#endif
+    .background(Color.expoSecondarySystemBackground)
     .cornerRadius(12)
   }
 
@@ -81,7 +79,7 @@ struct SettingsTabView: View {
         Image(systemName: "gearshape")
           .resizable()
           .frame(width: 56, height: 56)
-          .foregroundColor(.black.opacity(0.3))
+          .opacity(0.3)
 
         Text("Settings")
           .font(.title2)
@@ -118,9 +116,7 @@ struct SettingsTabView: View {
         }
         .padding()
       }
-#if !os(tvOS)
-      .background(Color(.systemGroupedBackground))
-#endif
+      .background(Color.expoSecondarySystemBackground)
       .cornerRadius(12)
     }
   }