+++ packages/expo-dev-menu/ios/SwiftUI/CustomItems.swift
@@ -26,9 +26,7 @@ struct CustomItems: View {
             }
             .padding()
           }
-          #if !os(tvOS)
-          .background(Color(.systemGroupedBackground))
-          #endif
+          .background(Color.expoSecondarySystemBackground)
           .clipShape(RoundedRectangle(cornerRadius: 12))
         }
       }